requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
anthropic>=0.40.0
aiosqlite>=0.20.0
asyncio
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
tenacity>=8.0.0
